Added API user getById and modified LogIn Security route

+++ New folder/document-review/document-review/src/main/java/com/example/document_review/service/UserService.java
@@ -6,7 +6,7 @@ import com.example.document_review.exception.EntityNotFoundException;
 
 public interface UserService {
     void register(UserDto userDto);
-    UserDto findById(long id);
+    UserDto findById(Integer id);
     User loginUser(UserDto userDto) throws EntityNotFoundException;
     User findByUsername(String username);
 }
